initial work on server
There's not too much to talk about here. The initial stuff has focused
on building things from the bottom up - starting with the racetrack
construction. There's some more details provided in the `sim` module,
which currently houses all of the work so far.

Nothing has yet been tested, although I'm reasonably confident that it
should be mostly correct; each component is quite simple - there's just
a bunch of pieces.

There's one final piece that still needs to be done on constructing the
racetrack, in the large function at the bottom of the file. I've left
that for myself to finish up, on the assumption that it'll probably be
easiest for everyone if I just quickly do that once I've gotten some
sleep.

The actual "making this into a server" part will come soon - the current
plan is to use rocket (https://rocket.rs/) for handling everything HTTP,
and pyo3 (https://pyo3.rs/) for embedding Python. The first-pass
solution shouldn't need to validate the Python or give it time limits,
but these can be solved by:
 1. Doing a pre-execution parsing stage to ensure there's no imports,
    "exec" expressions, or certain types of try-catch blocks, and
 2. Using unix alarm signals to trigger exceptions if it runs out of
    time