Finish racetrack construction (untested)
There were a couple shortcuts required -- the original plan for
constructing the finish line was going to involve an (unnecessarily)
complex algorithm for determining the line that passes through only the
finish-line tiles.

Instead of doing that, we're just requiring for now that the finish line
goes horizontally through the starting tile. A more complex solution can
be made later, if necessary.