shift movement bounds to something more reasonable
This is basically an adjustment so that cars take longer to go around
the track - the previous constants had a maximum speed of 10 tiles per
tick, with 100 ticks per second. That's 1000 tiles per second!!!

Our racetrack is no where near big enough for that to make sense, so it
seemed like the best solution would be to scale the numbers by the
number of ticks per second. That way, we can change the simulation rate
if need be.