remove `Copy` implementation on `Car`
This might seem counterproductive at first - let me explain.

In Rust, types implementing `Copy` can be implicitly copied - i.e.
something that was behind a reference can be implicitly turned into a
value if we'd otherwise *move* the data from it. This generally makes
sense - if we have an array of integers, we can use normal indexing and
it gets us the value.

This can bite you in the ass if you're not careful though. When the type
serving as your mutable state implements `Copy`, it's easily possible to
*accidentally* copy the value when you meant to mutate it. This is not
ideal.

So because the simulation keeps an inner `Car` as its mutable state,
I've decided it'll be safer to remove the `Copy` implementation - just
so nothing funky happens.