Add racetrack construction test, fix border angles
Tests are always nice to have -- and turns out there was a little bug in
the way that diagonal borders were being consructed. That bug *probably*
wasn't seriously impacting the collision detection (only racetrack tiles
with "outside" tiles to the left and above were impacted), but either
way it's gone now.