Add first-pass scripting langauge docs
This is available for whatever it's needed for in the front-end -- I'm
not sure where it goes, though, so I'm leaving that up to someone else.

There's a few changes coming in shortly that should change this, but I
figured it'd be good to get it up first before everything else.