Rewrite collision-detection code
The collision detection code may be incorrect. I'm not sure -- though I
belive it's mathematically sound. Unfortunately, it doesn't produce the
same results as the previous implementation -- and often not by a small
amount (± 2 units in some cases).

Because I don't know which version is correct -- and because this causes
the existing, custom script to not work, I've elected to leave this in a
separate branch for now.